Main.java Edits
If no values are given in the add/edits of Employee and Department, not
specified will now show.

In the case of Department, empty value adds would add a department which
could not be deleted. This removes the above issue as not specified will
always show if a null value is present. This would happen due to the
logic and syntax used to restrict the "please select" department which
has a id of 0 and cannot be deleted or edited.